// src/enums/UserMenuCommands.ts
/**
 * Enum con todos los comandos del menu de usuario
 */
export enum UserNormalUserComands {
  RoutesOptions = "Show options for Routes",
  ChallengeOptions = "Show options for Challenges",
  UserOptions = "Show options for Users and friends",
  GroupOptions = "Show options for Groups",
  Quit = "Quit",
}