// src/enums/UserChallengeCommands.ts
/**
 * Enum para todos los comandos de usuario de retos
 */
export enum UserChallengeCommands {
  ShowChallengeAlphabetically = "Show Challenges alphabetically",
  ShowChallengeReversedAlphabetically = "Show Challenges alphabetically reversed",
  ShowChallengeByKilometers = "Show Challenges order by number of kilometers",
  ShowChallengeByKilometersReversed = "Show Challenges order by number of kilometers reversed",
  ShowChallengeByTotalUsers = "Show Challenges order by number of TotalUsers",
  ShowChallengeByTotalUsersReversed = "Show Challenges order by number of TotalUsers reversed",
}